feat(icones): index icône en uint16_t + tables générées (firmware + tests)
Le set MDI (469) dépasse 255 : l'index d'icône passe de uint8_t à uint16_t sur
TOUTE la chaîne, sinon troncature silencieuse (ICON_GLYPHS[sym & 0xFF]).
- dashboard.{h,cpp} : IconState.symbol / Component.icon_symbol en uint16_t ;
  tables locales supprimées → icons_gen.h (généré) ; icon_symbol_index exposé (uint16_t).
- color.{h,cpp} : icon_resolve base_sym/out_sym/sym en uint16_t (non couvert par le plan,
  découvert en lisant les appelants — sinon tronque les états d'icône).
- platformio.ini : env native compile fonts/icons_gen.c (link des tables pour les tests).
- tests : icon_resolve/icon_parsed en uint16_t + assertions round-trip nom↔index
  (robustes au set) ; nouveau test_icon_symbol_index_legacy (compat des 23 noms).

189 tests natifs OK.

+++ src/color.cpp
@@ -20,9 +20,9 @@ bool led_is_lit(int32_t value, int32_t off_below) {
     return value >= off_below;
 }
 
-void icon_resolve(const IconState* st, int n, float value, uint8_t base_sym, uint32_t base_col,
-                  uint8_t* out_sym, uint32_t* out_col) {
-    uint8_t sym = base_sym; uint32_t col = base_col;
+void icon_resolve(const IconState* st, int n, float value, uint16_t base_sym, uint32_t base_col,
+                  uint16_t* out_sym, uint32_t* out_col) {
+    uint16_t sym = base_sym; uint32_t col = base_col;
     for (int i = 0; i < n; i++) {
         if (value < st[i].at) {
             if (st[i].has_symbol) sym = st[i].symbol;
